+++ app/src/main/java/com/example/stablestable/navigation/Navigation.kt
@@ -4,14 +4,16 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
+import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
 import com.example.stablestable.ui.home.HomeScreen
 import com.example.stablestable.ui.login.LoginScreen
-import com.example.stablestable.ui.stable.StableUsersScreen
 import com.example.stablestable.ui.horses.HorseProfileScreen
 import com.example.stablestable.ui.stable.horses.StableHorsesScreen
 import com.example.stablestable.ui.profile.UserProfileScreen
+import com.example.stablestable.ui.stable.riders.StableUsersScreen
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
@@ -53,7 +55,9 @@ fun SetupNavGraph(
             StableUsersScreen(
                 onUserClick = {
                         userId -> navController.navigate(Screen.UserProfileScreen.route.replace("{userId}", userId))
-                }
+                },
+                goToStable = { navController.navigate(Screen.StableUsers.route) },
+                goToHome = { navController.navigate(Screen.HomeScreen.route) }
             )
         }
 
@@ -64,27 +68,37 @@ fun SetupNavGraph(
             StableHorsesScreen(
                 onHorseClick = {
                         horseId -> navController.navigate(Screen.HorseProfileScreen.route.replace("{horseId}", horseId))
-                }
+                },
+                goToStable = { navController.navigate(Screen.StableUsers.route) },
+                goToHome = { navController.navigate(Screen.HomeScreen.route) }
             )
         }
 
         // My Profile Screen Route
         composable(
-            route = Screen.UserProfileScreen.route
+            route = Screen.UserProfileScreen.route,
+            arguments = listOf(navArgument("userId") { type = NavType.StringType })
         ) {backStackEntry ->
             UserProfileScreen(
-                backStackEntry.arguments?.getString("userId") ?: "",
+                userId = backStackEntry.arguments?.getString("userId") ?: "",
                 onHorseClick = {
                         horseId -> navController.navigate(Screen.HorseProfileScreen.route.replace("{horseId}", horseId))
-                }
+                },
+                goToStable = { navController.navigate(Screen.StableUsers.route) },
+                goToHome = { navController.navigate(Screen.HomeScreen.route) }
             )
         }
 
         // Horse Profile Screen Route
         composable(
-            route = Screen.HorseProfileScreen.route
+            route = Screen.HorseProfileScreen.route,
+                arguments = listOf(navArgument("horseId") { type = NavType.StringType })
         ) { backStackEntry ->
-            HorseProfileScreen(backStackEntry.arguments?.getString("horseId") ?: "")
+            HorseProfileScreen(
+                horseId = backStackEntry.arguments?.getString("horseId") ?: "",
+            goToStable = { navController.navigate(Screen.StableUsers.route) },
+            goToHome = { navController.navigate(Screen.HomeScreen.route) }
+            )
         }
     }
 }
