--- app/src/main/java/com/example/stablestable/ui/stable/riders/StableScreen.kt
@@ -1,13 +1,0 @@
-package com.example.stablestable.ui.stable.riders
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun StableScreen(
-) {
-    Column {
-        Text(text = "Stable Information Screen")
-    }
-}
