+++ app/src/main/java/com/example/stablestable/ui/stable/riders/StableUsersScreen.kt
@@ -1,17 +1,17 @@
-package com.example.stablestable.ui.home
+package com.example.stablestable.ui.stable.riders
 
 import androidx.compose.runtime.Composable
 import com.example.stablestable.components.CreateScaffold
 
 @Composable
-fun HomeScreen(
+fun StableUsersScreen(
     goToStable: () -> Unit,
-    goToHome: () -> Unit
-    //goToShifts: () -> Unit
-) {
+    goToHome: () -> Unit,
+    onUserClick: (String) -> Unit
+){
     CreateScaffold(
         content = {paddingValues ->
-        CreateHomeScreen(paddingValues)
+            StableUsersScreenContent(paddingValues, onUserClick)
         },
         goToHome = goToHome,
         goToStable = goToStable
